Made it so that you can repeat a for loop with retry.
I'm not entirely sure about this implementation, as it strikes me as touching too many lines of code for what is really a fairly simple thing. It's the cleanest solution I was able to find, though.

The block for a for loop is now made to have its own iseq type as that seems to be the only way to indicate its difference from a normal block (that there is a valid retry catcher for it). As such, most places where it checks if something is a block it now checks if it's a for-block as well.